refine export test helpers
